// src/generated/data-contracts.ts
export enum CurrencyEnum {
  USD = 'USD',
  ZAR = 'ZAR',
  AED = 'AED',
}

/** The country in which the business was incorporated */
export enum IncorporationCountryEnum {
  EG = 'EG',
  JO = 'JO',
  SA = 'SA',
  AE = 'AE',
  ES = 'ES',
  ZA = 'ZA',
}

/**
 * The credit score associated to the customer
 * @example 1
 */
export enum CreditScoreEnum {
  Value1 = 1,
  Value2 = 2,
  Value3 = 3,
  Value4 = 4,
  Value5 = 5,
  Value6 = 6,
  Value7 = 7,
  Value8 = 8,
}

export interface MetaData {
  /** Number of items */
  itemsCount?: number;
  /** Limit of items per page */
  limit?: number;
  /** Current page number */
  page?: number;
}

/**
 * Internal risk associated with the customer
 * @example "LOW"
 */
export enum InternalRisk {
  VERY_HIGH = 'VERY_HIGH',
  HIGH = 'HIGH',
  NEUTRAL = 'NEUTRAL',
  LOW = 'LOW',
}

export enum OrderByEnum {
  Asc = 'asc',
  Desc = 'desc',
}
// ...
export type LoanResponse = (NewLoan & object) & {
  loanOfficerUser?: LoanOfficerUser;
  /**
   * Unique loan identifier
   * @example "BtCSNcteA2oD226"
   */
  loanId?: string;
  /**
   * Loan current status
   * @example "Pending"
   */
  loanStatus?: LoanResponseLoanStatusEnum;
  /**
   * The date finance team transfers the funds to the client
   * @example "2023-05-06T00:00:00.000Z"
   */
  disbursementDate?: string;
  /**
   * The date on which the loan is due for final repayment
   * @example "2024-05-06T00:00:00.000Z"
   */
  maturityDate?: string;
  /** Currency */
  currency?: CurrencyEnum;
  /**
   * Date of which the loan is  actually paid off
   * @example "2024-12-06T00:00:00.000Z"
   */
  payoffDate?: string;
  /**
   * Date the loan was approved
   * @example "2023-12-06T00:00:00.000Z"
   */
  loanApprovalDate?: string;
  /**
   * How many times the loan has been deliquent
   * @example 0
   */
  timesDelinquent?: number;
  /**
   * The principal left to be paid
   * @example 0
   */
  netPayablePrincipal?: number;
  /**
   * Late payment fees for this loan
   * @example 0
   */
  netPayablePenalties?: number;
  /**
   * Total left sum of pricnipal, interest, penalties amount
   * @example 0
   */
  netPayableTotal?: number;
  /**
   * Total left sum of interest amount
   * @example 0
   */
  netPayableInterest?: number;
  /**
   * Maximum allowable disbursements including offset
   * @example 0
   */
  maxAllowableDisbursements?: number;
  /**
   * The total money disbursed to the customer
   * @example 0
   */
  totalDisbursedAmount?: number;
  /**
   * Total Principal paid by the customer
   * @example 0
   */
  paidPrincipalAmount?: number;
  /**
   * Total interest paid by the customer
   * @example 0
   */
  paidInterestAmount?: number;
  /**
   * Total penalties paid by the customer
   * @example 0
   */
  paidPenaltiesAmount?: number;
  /**
   * Late payment fees for this loan
   * @example 0
   */
  totalPenaltyAmount?: number;
  /**
   * Total offsets for the loan
   * @example 0
   */
  totalOffsetAmount?: number;
  /**
   * Total amount waived off
   * @example 0
   */
  waivedAmount?: number;
  /**
   * Penalty amount waived off
   * @example 0
   */
  waivedPenaltyAmount?: number;
  /**
   * Total interest adjusted
   * @example 0
   */
  adjustedInterestAmount?: number;
  /**
   * Total interest discounted
   * @example 0
   */
  allocatedDiscountInterestAmount?: number;
  /**
   * Principal Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  writtenOffPrincipalAmount?: number;
  /**
   * Interests Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  writtenOffInterestAmount?: number;
  /**
   * Total Amount Written off; this is the amount considered not recoverable
   * @example 0
   */
  totalWrittenOffAmount?: number;
  /**
   * This is the amount which was  restructured as part of a different loan
   * @example 0
   */
  restructuredAmount?: number;
  /**
   * The amount of interest
   * @example 1500
   */
  interestAmount?: number;
  /**
   * Date of next payment
   * @example "2023-05-06T00:00:00.000Z"
   */
  nextPaymentDate?: string;
  /**
   * Number of payments equal to 1 month
   * @example 12
   */
  numberOfPayments?: number;
  /**
   * Number of current payment
   * @example 0
   */
  currentPaymentNumber?: number;
  /**
   * Current delinquent amount of loan
   * @example 0
   */
  delinquentAmount?: number;
  /**
   * Origination fee
   * @example 0
   */
  originationFee?: number;
  /**
   * Origination fee amount
   * @example 0
   */
  originationFeeAmount?: number;
  /**
   * Origination fee amount
   * @example 0
   */
  totalChargesAmount?: number;
};
// ...
export interface NewPaymentLinkRequest {
  /**
   * Amount to be requested for payment
   * @example 200
   */
  amount: number;
  /**
   * Date when payment link expieres
   * @example "2000-12-21"
   */
  expirationDate?: string;
  /** Currency */
  currency: CurrencyEnum;
}

export interface NewLoan {
  /**
   * Opportunity ID in salesforce if loan came from it.
   * @example "00B7Q00000GWcU9UAL"
   */
  salesforceOpportunityId?: string;
  /**
   * The country in which the business was incorporated
   * @example "My Company Inc EG"
   */
  companyIdentifier: string;
  /**
   * The owner is the Sales team member managing the relationship with the client
   * @example "123e4567-e89b-12d3-a456-426614174000"
   */
  loanOfficerUserId: string;
  /**
   * Type of the loan
   * @example "RBF"
   */
  loanType: NewLoanLoanTypeEnum;
  /**
   * The number of days the payment can be late
   * @example 3
   */
  gracePeriodDays: number;
  /**
   * The total amount of time the client is requesting funding for.
   * @example 30
   */
  tenor?: number;
  /**
   * The percent of Penalty applied to this loan if a payment is late
   * @minLength 1
   * @maxLength 100
   * @example 5
   */
  penaltyPercent: number;
  /**
   * Type of penalties
   * @example "FIXED"
   */
  penaltyType: NewLoanPenaltyTypeEnum;
  /**
   * Method of allocation of interest
   * @example "INTEREST_UPFRONT"
   */
  interestAllocationType?: NewLoanInterestAllocationTypeEnum;
  /**
   * The monthly rate for IBF or fixed fee for RBF offered to the client for each loan .
   * @minLength 1
   * @maxLength 100
   * @example 15
   */
  monthlyRate: number;
  /** @example "monthly_fixed" */
  repaymentType?: NewLoanRepaymentTypeEnum;
  /**
   * The loan amount. This is the amount that is disbursed to the customer
   * @example 10000
   */
  principalAmount: number;
  /**
   * Total amount of invoices for which the customer requests for IBF loan
   * @example 10000
   */
  invoicesAmount?: number;
  /**
   * Rate of the amount of invoices we lend out as a loan
   * @minLength 1
   * @maxLength 100
   * @example 100
   */
  invoicesDiscountRate?: number;
  /**
   * First instalment due date
   * @example "2024-05-05T00:00:00.000Z"
   */
  firstInstalmentDate?: string;
  /**
   * The collection method used for the repayments
   * @example "BANK_TRANSFER"
   */
  collectionMethod?: string;
}

/** @example {"exampleOperation":{"href":"/example"}} */
export type HalLink = Record<
  string,
  {
    href?: string;
  }
>;
// ...
export interface OzowPaymentLinkRequest {
  /** Payment Link Request Id */
  id?: string;
  /** Loan_id for which payment link request is made */
  loanId?: string;
  /** Amount requested */
  amount?: number;
  /** Bank reference */
  bankReference?: string;
  /** Payment link expiry datetime */
  expiryDatetime?: Date;
  /** Payment link request datetime */
  createdAt?: Date;
  /** Ozow payment link url */
  url?: string;
  /** Error message if was not able to generate payment link */
  errorMessage?: string;
  /** Payment status */
  status?: string;
  /** Flag to indicate if reconciled on lms */
  isReconciled?: boolean;
  paymentRequest?: OzowPaymentRequest;
}

export interface ViewOzowPaymentLinkRequests {
  paymentLinkRequests?: OzowPaymentLinkRequest[];
  metaData?: MetaData;
  _links?: HalLink[];
  _embedded?: {
    loan?: LoanResponse;
  };
}
// ...
export interface CancelLoanRequest {
  note?: string;
}
// ...
export enum NewLoanLoanTypeEnum {
  RBF = 'RBF',
  IBF = 'IBF',
  IBF2 = 'IBF2',
}

/**
 * Type of penalties
 * @example "FIXED"
 */
export enum NewLoanPenaltyTypeEnum {
  FIXED = 'FIXED',
  VARIABLE = 'VARIABLE',
}

/**
 * Method of allocation of interest
 * @example "INTEREST_UPFRONT"
 */
export enum NewLoanInterestAllocationTypeEnum {
  INTEREST_UPFRONT = 'INTEREST_UPFRONT',
  SPLIT_EQUALLY = 'SPLIT_EQUALLY',
}

/** @example "monthly_fixed" */
export enum NewLoanRepaymentTypeEnum {
  MonthlyFixed = 'monthly_fixed',
  WeeklyFixed = 'weekly_fixed',
  EndOfTenor = 'end_of_tenor',
}
// ...
export enum GetCompanyLoansParams1IncludeEnum {
  SalesforceOpportunityId = 'salesforceOpportunityId',
  LoanOfficerUserId = 'loanOfficerUserId',
  GracePeriodDays = 'gracePeriodDays',
  PenaltyPercent = 'penaltyPercent',
  PenaltyType = 'penaltyType',
  LoanStatus = 'loanStatus',
  DisbursementDate = 'disbursementDate',
  MaturityDate = 'maturityDate',
  PayoffDate = 'payoffDate',
  LoanApprovalDate = 'loanApprovalDate',
  TimesDelinquent = 'timesDelinquent',
  TotalDisbursedAmount = 'totalDisbursedAmount',
  PaidPrincipalAmount = 'paidPrincipalAmount',
  PaidInterestAmount = 'paidInterestAmount',
  PaidPenaltiesAmount = 'paidPenaltiesAmount',
  TotalPenaltyAmount = 'totalPenaltyAmount',
  CreatedAt = 'createdAt',
}
// ...
export interface GetLoansParams {
  /** In which order will the data be sorted (ascending or descending) */
  orderBy?: OrderByEnum;
  /** Search items by partial matching of search param */
  search?: string;
  /** Number of items in the response */
  limit?: number;
  /** Page number */
  page?: number;
  /** Company field by which the records will be sorted */
  sortBy?: SortByEnum2;
  /** Loan fields that can be included */
  include?: GetCompanyLoansParams1IncludeEnum[];
}

/** Company field by which the records will be sorted */
export enum SortByEnum2 {
  CompanyIdentifier = 'companyIdentifier',
  LoanOfficerUserId = 'loanOfficerUserId',
  LoanType = 'loanType',
  GracePeriodDays = 'gracePeriodDays',
  Tenor = 'tenor',
  PenaltyPercent = 'penaltyPercent',
  PenaltyType = 'penaltyType',
  MonthlyRate = 'monthlyRate',
  LoanStatus = 'loanStatus',
  DisbursementDate = 'disbursementDate',
  MaturityDate = 'maturityDate',
  PayoffDate = 'payoffDate',
  LoanApprovalDate = 'loanApprovalDate',
  TimesDelinquent = 'timesDelinquent',
  PrincipalAmount = 'principalAmount',
  InterestAmount = 'interestAmount',
  TotalDisbursedAmount = 'totalDisbursedAmount',
  PaidPrincipalAmount = 'paidPrincipalAmount',
  PaidInterestAmount = 'paidInterestAmount',
  PaidPenaltiesAmount = 'paidPenaltiesAmount',
  TotalPenaltyAmount = 'totalPenaltyAmount',
  CreatedAt = 'createdAt',
}
// ...
export interface SetLoanDefaultReviewStatusPayload {
  /**
   * Note
   * @example "Some note"
   */
  note?: string;
}

export type SetLoanDefaultReviewStatusData = (LoanResponse & object) & {
  _links?: HalLink[];
};

export interface UnsetLoanDefaultReviewStatusPayload {
  /**
   * Note
   * @example "Some note"
   */
  note?: string;
}

export type UnsetLoanDefaultReviewStatusData = (LoanResponse & object) & {
  _links?: HalLink[];
};

export interface WriteOffLoanPayload {
  /**
   * Note
   * @example "Some note"
   */
  note?: string;
}
// ...
export type GetPaymentLinkRequestsData = ViewOzowPaymentLinkRequests;